User story 4 roles

+++ app/controllers/wikis_controller.rb
@@ -1,22 +1,24 @@
 class WikisController < ApplicationController
-  before_action :authenticate_user!
 
   def index
     @wikis = Wiki.all
+    authorize @wikis
   end
 
   def show
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
   end
 
   def new
     @wiki = Wiki.new
+    authorize @wiki
   end
 
   def create
     @wiki = Wiki.new(wiki_params)
     @wiki.user = current_user
-
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was saved."
@@ -29,11 +31,14 @@ class WikisController < ApplicationController
 
   def edit
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
+
   end
 
   def update
     @wiki = Wiki.find(params[:id])
     @wiki.assign_attributes(wiki_params)
+    authorize @wiki
 
     if @wiki.save
       flash[:notice] = "Wiki was updated."
@@ -46,17 +51,25 @@ class WikisController < ApplicationController
 
   def destroy
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
 
     if @wiki.destroy
       flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
-      redirect_to wikis_path
+      redirect_to @wiki
     else
       flash.now[:alert] = "There was an error deleting the wiki."
       render :show
     end
   end
-end
 
-def wiki_params
-  params.require(:wiki).permit(:title, :body)
+  private
+
+  def wiki_params
+    params.require(:wiki).permit(:title, :body)
+  end
+
+  def user_not_authorized
+    flash[:alert] = "You must be authorized to do this."
+    redirect_to @wiki
+  end
 end
